docker/requirements.txt
h3==3.7.4
folium==0.13.0
geopandas==0.10.2
rioxarray==0.13.2
Shapely==1.8.0
pyroscope-io==0.8.0
Rtree==1.0.1